pyproject.toml
[project]
name = "tanzpalast"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "pyyaml",
    "playwright",
    "pytest",
    "pytest-playwright",
    "pytest-cov>=7.1.0",
]
